# .agents/resources/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_import_workflow_api_v1_workflows_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Override"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_workflow_api_v1_workflows_import_post",
        "type": "object"
      },
...
      "ColumnRefEdge": {
        "description": "Edge that maps a named output column to a named input.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "source_output": {
            "title": "Source Output",
            "type": "string"
          },
          "target_input": {
            "title": "Target Input",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "column_ref",
            "default": "column_ref",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "source_output",
          "target_input"
        ],
        "title": "ColumnRefEdge",
        "type": "object"
      },
...
      "GraphState": {
        "description": "Complete graph state with nodes and edges.",
        "properties": {
          "edges": {
            "items": {
              "discriminator": {
                "mapping": {
                  "column_ref": "#/components/schemas/ColumnRefEdge",
                  "positional": "#/components/schemas/PositionalEdge"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ColumnRefEdge"
                },
                {
                  "$ref": "#/components/schemas/PositionalEdge"
                }
              ]
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeState"
            },
            "title": "Nodes",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "GraphState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MissingPackage": {
        "description": "Package/version-level load issue for a workflow.",
        "properties": {
          "affected_nodes": {
            "items": {
              "type": "string"
            },
            "title": "Affected Nodes",
            "type": "array"
          },
          "installed_versions": {
            "items": {
              "type": "string"
            },
            "title": "Installed Versions",
            "type": "array"
          },
          "package_name": {
            "title": "Package Name",
            "type": "string"
          },
          "required_version": {
            "title": "Required Version",
            "type": "string"
          }
        },
        "required": [
          "package_name",
          "required_version"
        ],
        "title": "MissingPackage",
        "type": "object"
      },
      "MissingTool": {
        "description": "Node/tool-level load issue for a workflow.",
        "properties": {
          "installed_versions": {
            "items": {
              "type": "string"
            },
            "title": "Installed Versions",
            "type": "array"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "package_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Name"
          },
          "required_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Version"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "tool_name"
        ],
        "title": "MissingTool",
        "type": "object"
      },
...
      "NodeState": {
        "description": "A single node in the processing graph.",
        "properties": {
          "collapsed": {
            "default": false,
            "title": "Collapsed",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_templates": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Output Templates",
            "type": "object"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "position": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Position",
            "type": "array"
          },
          "published_inputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedInput"
            },
            "title": "Published Inputs",
            "type": "array"
          },
          "published_outputs": {
            "items": {
              "$ref": "#/components/schemas/PublishedOutput"
            },
            "title": "Published Outputs",
            "type": "array"
          },
          "resources": {
            "additionalProperties": true,
            "title": "Resources",
            "type": "object"
          },
          "sub_workflow": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphState"
              },
              {
                "type": "null"
              }
            ]
          },
          "sub_workflow_readonly_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Workflow Readonly Reason"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tool_name",
          "position",
          "parameters"
        ],
        "title": "NodeState",
        "type": "object"
      },
...
      "PositionalEdge": {
        "description": "Edge that connects nodes by positional index.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "positional_index": {
            "title": "Positional Index",
            "type": "integer"
          },
          "source_node": {
            "title": "Source Node",
            "type": "string"
          },
          "target_node": {
            "title": "Target Node",
            "type": "string"
          },
          "type": {
            "const": "positional",
            "default": "positional",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_node",
          "target_node",
          "positional_index"
        ],
        "title": "PositionalEdge",
        "type": "object"
      },
      "PublishedInput": {
        "description": "An internal sub-workflow field exposed as an outer input pin.",
        "properties": {
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "internal_field": {
            "title": "Internal Field",
            "type": "string"
          },
          "internal_node_id": {
            "title": "Internal Node Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "parameter",
              "input"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          }
        },
        "required": [
          "name",
          "internal_node_id",
          "internal_field",
          "kind"
        ],
        "title": "PublishedInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowCreate": {
        "description": "Request body for POST /workflows.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "name"
        ],
        "title": "WorkflowCreate",
        "type": "object"
      },
...
      "WorkflowFile": {
        "description": "Workflow file response returned to the frontend.",
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/GraphState"
          },
          "gui": {
            "additionalProperties": true,
            "title": "Gui",
            "type": "object"
          },
          "info": {
            "$ref": "#/components/schemas/WorkflowInfo"
          },
          "missing_packages": {
            "items": {
              "$ref": "#/components/schemas/MissingPackage"
            },
            "title": "Missing Packages",
            "type": "array"
          },
          "missing_tools": {
            "items": {
              "$ref": "#/components/schemas/MissingTool"
            },
            "title": "Missing Tools",
            "type": "array"
          }
        },
        "required": [
          "info",
          "graph"
        ],
        "title": "WorkflowFile",
        "type": "object"
      },
      "WorkflowImportResponse": {
        "description": "Workflow import success response.",
        "properties": {
          "info": {
            "$ref": "#/components/schemas/WorkflowInfo"
          },
          "missing_packages": {
            "items": {
              "$ref": "#/components/schemas/MissingPackage"
            },
            "title": "Missing Packages",
            "type": "array"
          },
          "missing_tools": {
            "items": {
              "$ref": "#/components/schemas/MissingTool"
            },
            "title": "Missing Tools",
            "type": "array"
          }
        },
        "required": [
          "info"
        ],
        "title": "WorkflowImportResponse",
        "type": "object"
      },
      "WorkflowInfo": {
        "description": "Workflow list item returned by GET /workflows.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "last_modified": {
            "title": "Last Modified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "name",
          "display_name",
          "path",
          "last_modified"
        ],
        "title": "WorkflowInfo",
        "type": "object"
      },
      "WorkflowSaveBody": {
        "description": "Request body for PUT /workflows/{name}.",
        "properties": {
          "draft_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphState"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "title": "WorkflowSaveBody",
        "type": "object"
      },
      "WorkflowUpdate": {
        "description": "Request body for PATCH /workflows/{name}.",
        "properties": {
          "action": {
            "enum": [
              "update",
              "duplicate"
            ],
            "title": "Action",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "new_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Name"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          }
        },
        "required": [
          "action"
        ],
        "title": "WorkflowUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BioImageFlow Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workflows": {
      "get": {
        "operationId": "list_workflows_api_v1_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowInfo"
                  },
                  "title": "Response List Workflows Api V1 Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "create_workflow_api_v1_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/import": {
      "post": {
        "operationId": "import_workflow_api_v1_workflows_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_workflow_api_v1_workflows_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{name}": {
      "delete": {
        "operationId": "delete_workflow_api_v1_workflows__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Workflow Api V1 Workflows  Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workflow",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "get_workflow_api_v1_workflows__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "patch": {
        "operationId": "patch_workflow_api_v1_workflows__name__patch",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Workflow",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "operationId": "save_workflow_api_v1_workflows__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSaveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{name}/export": {
      "post": {
        "operationId": "export_workflow_api_v1_workflows__name__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/v1/workflows/{name}/rebind-versions": {
      "post": {
        "operationId": "rebind_workflow_versions_api_v1_workflows__name__rebind_versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebind Workflow Versions",
        "tags": [
          "workflows"
        ]
      }
    }
  }
}